backend/requirements.txt
# TODO: remove asyncpg-related libraries
python-dotenv
fastapi
uvicorn
sqlmodel
asyncpg
psycopg_binary
langchain
langchain-ollama
langchain-cloudflare
langchain-postgres
langgraph-checkpoint-postgres
greenlet